requirements.txt
alembic==1.17.1
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.60.1
imbalanced-learn==0.14.0
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
numba==0.62.1
numpy==2.2.6
openpyxl==3.1.5
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scikit-plot==0.3.7
scipy==1.15.3
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.44
threadpoolctl==3.6.0
tomli==2.3.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
xgboost==3.1.1
xlrd==2.0.2
keras==3.12.0
tensorflow==2.20.0